[FIND-215] Handle null files
Turns out we look for a file called null if nothing is configured. Trap
and prevent this case.

[rev: will.byrne]

+++ webapp/idol/src/main/java/com/hp/autonomy/frontend/find/idol/customisations/CustomisationServiceImpl.java
@@ -75,6 +75,10 @@ public class CustomisationServiceImpl implements CustomisationService {
 
     @Override
     public boolean deleteAsset(final AssetType assetType, final String name) {
+        if (name == null) {
+            return false;
+        }
+
         return getAsset(assetType, name).delete();
     }
